mito-ai: create utilities
